fix a bunch of errors on python3

+++ localgraphclustering/graph_lib/wrappers/python/list_to_CSR.py
@@ -12,8 +12,8 @@ def list_to_CSR(filename):
     first_line = f.readline().strip()
     import re
     first_line = re.split(" ", first_line)
-    n = long(first_line[0])
-    m = long(first_line[2])
+    n = int(first_line[0])
+    m = int(first_line[2])
     ei = []
     ej = []
     w = []
